Memory optimizations and various fixes (#87)
* First round of memory fixups

* More performant echo filter using a ring buffer

* Significant improvements in GC pressure issues

* Fix crash on playback with some effects

* Fix s3m/mod pattern playback and pcm sampling

+++ internal/format/xm/playback/effect/intf/intf.go
@@ -9,5 +9,5 @@ type XM interface {
 	SetTempo(int) error
 	DecreaseTempo(int) error
 	IncreaseTempo(int) error
-	SetEnvelopePosition(uint8)
+	SetEnvelopePosition(int)
 }
